Add a classDiagramKey attribute to the diagram entity.
In the case where the diagram entity is an object diagram, the classDiagramkey is the key for the class diagram instantiated.

When the getDiagram service is called, if the diagram found is an object diagram, the service also add the classes from the class diagram instantiated to the object diagram found.